got the android pose pipeline running end to end on a pixel 7. added android.permission.CAMERA + camera uses-feature to AndroidManifest because the visioncamera expo plugin's auto-injection didn't survive the bare workflow prebuild — without it requestPermission silently returned false and the camera never mounted. fixed JSI bridge crash "Cannot convert Java type 'class java.lang.Float' to jsi::Value" by upcasting NormalizedLandmark x/y/z/visibility to Double in PoseLandmarkerFrameProcessorPlugin since visioncamera v4 only accepts Double, not java.lang.Float. added a Platform.OS branch in VisionService.ts so android applies y: 1 - lm.x to compensate for cameraX delivering the front camera buffer in the opposite vertical orientation from iOS, plus a separate non-swapped MEDIAPIPE_TO_COCO_ANDROID label table because the android buffer isn't pre-mirrored like iOS so the iOS-calibrated L/R label swap inverted body sides. iOS path is byte-identical — same code, same table contents, just renamed to MEDIAPIPE_TO_COCO_IOS. commented out the debug angle overlay in exercise-session.tsx for device testing. R3 clinical L/R correctness verified on a pixel 7 with a left-side hip abductor.

+++ front-end/AI-PEER/android/app/src/main/java/com/anonymous/aipeer/poselandmarker/PoseLandmarkerFrameProcessorPlugin.kt
@@ -174,20 +174,21 @@ class PoseLandmarkerFrameProcessorPlugin(
         if (landmarks.size < 33) return null
 
         // ── Build JS-bridge-friendly output ────────────────────────────────
-        // Must be List<Map<String, Any>> with String keys and primitive Float
-        // values (NOT Double) to match the iOS dict shape exactly. Java
-        // ArrayList → JS Array, HashMap → JS Object via the JSI bridge.
-        // NormalizedLandmark.x/y/z/visibility return primitive float, so
-        // putting them into HashMap<String, Any> auto-boxes to java.lang.Float.
+        // Must be List<Map<String, Any>> with String keys and Double values.
+        // VisionCamera v4's JSI bridge on Android does NOT support java.lang.Float
+        // ("Cannot convert Java type 'class java.lang.Float' to jsi::Value!") so
+        // we explicitly upcast each primitive float to Double before boxing into
+        // the map. JS receives plain `number` either way, so iOS parity is
+        // preserved at the contract level.
         val output = ArrayList<Map<String, Any>>(landmarks.size)
         for (lm in landmarks) {
             val dict = HashMap<String, Any>(4)
-            dict["x"] = lm.x()
-            dict["y"] = lm.y()
-            dict["z"] = lm.z()
+            dict["x"] = lm.x().toDouble()
+            dict["y"] = lm.y().toDouble()
+            dict["z"] = lm.z().toDouble()
             // NormalizedLandmark.visibility() returns java.util.Optional<Float>.
             // Match iOS: include the key only when visibility is present.
-            lm.visibility().ifPresent { dict["visibility"] = it }
+            lm.visibility().ifPresent { dict["visibility"] = it.toDouble() }
             output.add(dict)
         }
         return output
